requirements.txt
azure-storage-blob==2.1.0
python-dotenv